fix realm users change. introducing new script
