+++ packages/matrix/package.json
@@ -33,6 +33,7 @@
     "register-bot-user": "MATRIX_USERNAME=aibot MATRIX_PASSWORD=pass ts-node --transpileOnly ./scripts/register-test-user.ts",
     "register-submission-bot": "MATRIX_IS_ADMIN=TRUE MATRIX_USERNAME=submissionbot MATRIX_PASSWORD=password ts-node --transpileOnly ./scripts/register-test-user.ts",
     "register-realm-user": "ts-node --transpileOnly ./scripts/register-realm-user.ts",
+    "register-realm-user-using-api": "ts-node --transpileOnly ./scripts/register-realm-user-using-api.ts",
     "register-realm-bot": "ts-node --transpileOnly ./scripts/register-realm-bot.ts",
     "register-test-user": "MATRIX_USERNAME=user MATRIX_PASSWORD=password ts-node --transpileOnly ./scripts/register-test-user.ts",
     "register-skills-writer": "MATRIX_USERNAME=skills_writer MATRIX_PASSWORD=password ts-node --transpileOnly ./scripts/register-test-user.ts",
