+++ packages/matrix/scripts/register-realm-bot.ts
@@ -1,4 +1,4 @@
-import { registerRealmUser } from './register-realm-user';
+import { registerRealmUser } from './register-realm-user-using-api';
 
 const realmServerURL = process.env.REALM_SERVER_URL || 'http://localhost:4201';
